+++ specs/ft_substr_specs.c
@@ -3,20 +3,20 @@
 
 char *ft_substr(char const *s, unsigned int start, size_t len);
 
-void display_result(char *original, char *new, int start, int len, int succes)
+void display_result(char *original, char *new, int start, size_t len, int succes)
 {
 	printf("--------------\n");
 	(succes) ? printf("\033[0;32mOK!\n") : printf("\033[0;31mKO!\n");
 	printf("Original = \"%s\"\n     New = \"%s\"\n", original, new);
 	printf("start index = %d\n", start);
-	printf("length max = %d actual length %d\n", len, ft_strlen(new));
+	printf("length max = %zu actual length %zu\n", len, ft_strlen(new));
 	printf("\033[0m");
 }
 
 void specs_ft_substr(int *succes, int *failure)
 {
 	int index;
-	int len;
+	size_t len;
 	char strs[4][14] = {"adore vraiment", "", " ", "\n"} ;
 	char *ptn_result;
 	for(index = 0; strs[0][index]; index++)
